+++ src/components/ButtonText/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
 import Button from 'src/components/Button';
-import withText from 'src/components/Button/withText';
+import withText from './withText';
 
 export default withText(props => <Button className="button-text" {...props} />);
