+++ src/components/ButtonText/withText.js
@@ -3,22 +3,26 @@ import styled from 'styled-components';
 export default Component => styled(Component)`
     ${({
         theme: {
-            secondary: { base, onBase, light, dark },
+            secondary: {
+                base, onBase, dark, lightOpacity,
+            },
         },
     }) => (`
             box-shadow: none;
             background-color: ${onBase};
             color: ${base};
             &:hover {
-                background-color: ${light};
+                background-color: ${lightOpacity};
                 color: ${dark};
+                box-shadow: none;
             }
             &:focus {
-                background-color: ${light};
+                background-color: ${lightOpacity};
                 color: ${dark};
+                box-shadow: none;
             }
             & .ripple {
-                background: green;
+
             }
             `)};
 `;
